Added support for a 'Dictionary with user-defined keys' in JSON schemas.
This allows us to model dictionaries in which the keys
aren't explicitly set by the schema. This idea will be
used in an upcoming MNX feature (lyric lines).

This change updates all parts of the system to handle
this: the admin, the JSONSchema generation, the
annotated example documents and the basic docs pages.