Changed JSON Schema generation to use defs for all non-native object types.
Previously, we only used defs for objects that appear in more
than one place. Now, we use defs for everything. The goal is
to remove an unnecessary special case and be consistent in output.
This should also make future schema diffs more easily readable,
since the changes will be more self-contained.