Added support for global JSON attributes.
If a JSONObject exists with the name '__globalattrs__', it will be
special-cased as follows:

* Its attributes will be included in all JSONObjects of type
OBJECT_TYPE_DICT -- both in the generated JSON Schema and in the
individual JSONObjects' docs pages.

* It won't have its own docs page (because its attributes are
included in all other objects' docs pages).

This effectively adds support for attributes such as 'id' that
we want to 'include' into every single object, defined in a single
place.